backend/requirements.txt
psycopg2-binary   # PostgreSQL driver
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose